Added and fixed stuff...

+++ src/fi/tamk/anpro/WeaponSwarm.java
@@ -39,7 +39,7 @@ public class WeaponSwarm extends AbstractWeapon
      * @param int Ampujan Y-koordinaatti
      */
 	@Override
-	public final void activate(int _targetX, int _targetY, float _startX, float _startY)
+	public final void activate(float _targetX, float _targetY, float _startX, float _startY)
 	{
 		// Soitetaan ‰‰ni
 		SoundManager.playSound(3, 1);
